chore: Rename `ReanimatedSlider.ts` to `Slider.ts`
